Refactor OfferingsPaywalls into VM
